requirements.txt
Django==2.2.10
djangorestframework==3.12.2
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
